integration

+++ src/software/jetson_nano/primitive_executor.cpp
@@ -13,8 +13,7 @@ PrimitiveExecutor::PrimitiveExecutor(const double time_step,
                                      const RobotId robot_id)
     : current_primitive_(),
       robot_constants_(robot_constants),
-      hrvo_simulator_(static_cast<float>(time_step), robot_constants,
-                      friendly_team_colour),
+      hrvo_simulator_(time_step, robot_constants, friendly_team_colour),
       time_step_(time_step),
       curr_orientation_(Angle::zero()),
       robot_id_(robot_id)
@@ -78,7 +77,8 @@ AngularVelocity PrimitiveExecutor::getTargetAngularVelocity(
 
     double max_angular_speed =
         static_cast<double>(robot_constants_.robot_max_ang_speed_rad_per_s);
-    double next_angular_speed = std::min(max_angular_speed, deceleration_angular_speed);
+    double next_angular_speed
+    = std::min(max_angular_speed, deceleration_angular_speed);
 
     const double signed_delta_orientation =
         (dest_orientation - curr_orientation_).clamp().toRadians();
